requirements.txt
numpy
scipy
nltk

python-jsonrpc
pymongo
pika
python-dateutil
lxml
newspaper
python-jsonrpc
pymongo
pika
redis
requests
sklearn
tensorflow
pandas
watchdog